A redundant .config change
I forget to enable network manually at guest as described at tiny-emu's
readme, I thought that I have to add `vtun` to the build to enable
guest's network function. But what is weired is that even I set choice
of `vtun` to NO, it still remains choices of its dependencies YES,
doesn't it suppose to back to NOT SET and leave only changed hash?